Tippett plots without extrapolation + type II plots (#66)
* Tippett plots without extrapolation

Modify the  Tippett plot function so it only shows values that are present in the validation data, in accordance with Section C.1 of the Consensus on validation of forensic voice comparison (https://doi.org/10.1016/j.scijus.2021.02.002): "Note that the curves do not extend to a y value of zero as they are representations of the empirical cumulative probability distribution, hence the lowest y value corresponds to 1/N where N is the number of same-speaker or different-speaker input pairs (for these illustrative data Ns = 50 and Nd = 200). Note also that the x values of the curves are not extrapolated beyond the values of the actual validation results obtained."

* Allow type II Tippett plots

Modify the Tippett plot function to allow Tippett plots of the same type described in the Consensus on validation of forensic voice comparison, Appendix C.1. Previous behavior is mantained with default option of the new parameter.